[FIX] not raise error for modifying ddt from sale order - but pass

+++ l10n_it_ddt_mobile/models/stock_picking_package_preparation.py
@@ -42,24 +42,17 @@ class StockPickingPackagePreparation(models.Model):
                 if line.move_id and line.product_id:
                     if line.move_id.picking_id:
                         picking = line.move_id.picking_id
-                        if picking.group_id:
-                            raise exceptions.Warning(
-                                _('Impossible to modify picking \
-                                  created from sale order.'))
-                        #TODO picking.group_id is not changed
-                        if line.move_id.product_uom_qty != line.product_uom_qty:
+                        if line.move_id.product_uom_qty != \
+                                line.product_uom_qty and not picking.group_id:
                             picking.action_cancel()
                             line.move_id.product_uom_qty = line.product_uom_qty
                             line.move_id.product_id = line.product_id
                             picking.force_assign()
 
             for picking in package.picking_ids:
-                if picking.group_id:
-                    raise exceptions.Warning(
-                        _('Impossible to modify picking \
-                          created from sale order.'))
                 for move in picking.move_lines:
-                    if move not in package.line_ids.mapped('move_id'):
+                    if move not in package.line_ids.mapped('move_id') and not \
+                            picking.group_id:
                         picking.action_cancel()
                         move.unlink()
                         picking.force_assign()
